refactor: move DEPARTMENTS and US-STATES to data
